requirements.txt
Flask==1.1.2
google-cloud-storage==1.29.0
gunicorn==20.0.4; python_version > '3.0'